Refactored network class, changes to input format
Input nodes now receive a value of 0 for empty values, 1 for
false values and 2 for positive values. Network class now contains
a private data-structure to enforce DRY and make code cleaner.
Overloaded index operator [] allows the output layer to be accessed
as if it is stored in the last index of the m_nodes vector, while
improving computational efficency by limiting move operations when
inserting new layers into the network.